+++ server/util_test.go
@@ -39,7 +39,7 @@ func preRunETestImpl(cmd *cobra.Command, args []string) error {
 
 func TestInterceptConfigsPreRunHandlerCreatesConfigFilesWhenMissing(t *testing.T) {
 	tempDir := t.TempDir()
-	cmd := server.StartCmd(server.StartCmdOptions{AppCreator: nil, DefaultNodeHome: "/foobar"})
+	cmd := server.StartCmd(nil, "/foobar")
 	if err := cmd.Flags().Set(flags.FlagHome, tempDir); err != nil {
 		t.Fatalf("Could not set home flag [%T] %v", err, err)
 	}
@@ -115,7 +115,7 @@ func TestInterceptConfigsPreRunHandlerReadsConfigToml(t *testing.T) {
 		t.Fatalf("Failed closing config.toml: %v", err)
 	}
 
-	cmd := server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: "/foobar"})
+	cmd := server.StartCmd(nil, "/foobar")
 	if err := cmd.Flags().Set(flags.FlagHome, tempDir); err != nil {
 		t.Fatalf("Could not set home flag [%T] %v", err, err)
 	}
@@ -155,7 +155,7 @@ func TestInterceptConfigsPreRunHandlerReadsAppToml(t *testing.T) {
 	if err := writer.Close(); err != nil {
 		t.Fatalf("Failed closing app.toml: %v", err)
 	}
-	cmd := server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: tempDir})
+	cmd := server.StartCmd(nil, tempDir)
 
 	cmd.PreRunE = preRunETestImpl
 
@@ -174,7 +174,7 @@ func TestInterceptConfigsPreRunHandlerReadsAppToml(t *testing.T) {
 func TestInterceptConfigsPreRunHandlerReadsFlags(t *testing.T) {
 	const testAddr = "tcp://127.1.2.3:12345"
 	tempDir := t.TempDir()
-	cmd := server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: "/foobar"})
+	cmd := server.StartCmd(nil, "/foobar")
 
 	if err := cmd.Flags().Set(flags.FlagHome, tempDir); err != nil {
 		t.Fatalf("Could not set home flag [%T] %v", err, err)
@@ -202,7 +202,7 @@ func TestInterceptConfigsPreRunHandlerReadsFlags(t *testing.T) {
 func TestInterceptConfigsPreRunHandlerReadsEnvVars(t *testing.T) {
 	const testAddr = "tcp://127.1.2.3:12345"
 	tempDir := t.TempDir()
-	cmd := server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: "/foobar"})
+	cmd := server.StartCmd(nil, "/foobar")
 	if err := cmd.Flags().Set(flags.FlagHome, tempDir); err != nil {
 		t.Fatalf("Could not set home flag [%T] %v", err, err)
 	}
@@ -288,7 +288,7 @@ func newPrecedenceCommon(t *testing.T) precedenceCommon {
 	})
 
 	// Set up the command object that is used in this test
-	retval.cmd = server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: tempDir})
+	retval.cmd = server.StartCmd(nil, tempDir)
 	retval.cmd.PreRunE = preRunETestImpl
 
 	return retval
@@ -395,7 +395,7 @@ func TestInterceptConfigsWithBadPermissions(t *testing.T) {
 	if err := os.Mkdir(subDir, 0o600); err != nil {
 		t.Fatalf("Failed to create sub directory: %v", err)
 	}
-	cmd := server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: "/foobar"})
+	cmd := server.StartCmd(nil, "/foobar")
 	if err := cmd.Flags().Set(flags.FlagHome, subDir); err != nil {
 		t.Fatalf("Could not set home flag [%T] %v", err, err)
 	}
@@ -432,7 +432,7 @@ func TestEmptyMinGasPrices(t *testing.T) {
 	config.WriteConfigFile(appCfgTempFilePath, appConf)
 
 	// Run StartCmd.
-	cmd = server.StartCmd(StartCmdOptions{AppCreator: nil, DefaultNodeHome: tempDir})
+	cmd = server.StartCmd(nil, tempDir)
 	cmd.PreRunE = func(cmd *cobra.Command, _ []string) error {
 		ctx, err := server.InterceptConfigsAndCreateContext(cmd, "", nil, cmtcfg.DefaultConfig())
 		if err != nil {